add compare_line_star
