+++ braincell/_integrators_implicit.py
@@ -23,17 +23,19 @@ import jax
 import jax.numpy as jnp
 from scipy.integrate import solve_ivp
 
-from ._integrators_util import apply_standard_solver_step
+
+from ._integrators_util import apply_standard_solver_step, _jacrev_last_dim
 from ._misc import set_module_as
 from ._protocol import DiffEqModule
-
+from ._integrators_runge_kutta import ralston4_step, rk4_step
+from ._integrators_exp_euler import _exponential_euler, exp_euler_step
 __all__ = [
     'implicit_euler_step',
     'splitting_step',
 ]
 
 
-def _newton_method(f, y0, t, dt, modified=False, tol=1e-5, max_iter=100, order=2, args=()):
+def _newton_method(f, y0, t, dt, args=(), modified=False, tol=1e-5, max_iter=100, order=2):
     r"""
     Newton's method for solving the implicit equations arising from the Crank - Nicolson method for ordinary differential equations (ODEs).
 
@@ -65,6 +67,7 @@ def _newton_method(f, y0, t, dt, modified=False, tol=1e-5, max_iter=100, order=2
     """
 
     def g(t, y, *args):
+        #jax.debug.print("arg = {a}", a = args)
         if order == 1:
             return y - y0 - dt * f(t + dt, y, *args)[0]
         elif order == 2:
@@ -75,10 +78,10 @@ def _newton_method(f, y0, t, dt, modified=False, tol=1e-5, max_iter=100, order=2
     def cond_fun(carry):
         i, _, cond = carry
         # condition = u.math.logical_or(u.math.linalg.norm(A) < tol, u.math.linalg.norm(df) < tol)
-        return u.math.logical_and(i < max_iter, u.math.logical_not(cond))
+        return u.math.logical_and(i < max_iter,cond)
 
     def body_fun(carry):
-        i, y1, _, _ = carry
+        i, y1, _= carry
         A, df = brainstate.augment.jacfwd(lambda y: g(t, y, *args), return_value=True, has_aux=False)(y1)
         # df: [n_neuron, n_compartment, M]
         # A: [n_neuron, n_compartment, M, M]
@@ -107,10 +110,8 @@ def _newton_method(f, y0, t, dt, modified=False, tol=1e-5, max_iter=100, order=2
     else:
         n, result, _, df = jax.lax.while_loop(cond_fun, body_fun_modified, init_carry)
     '''
-    n, result, _, _ = jax.lax.while_loop(cond_fun, body_fun, init_carry)
+    n, result, _ = jax.lax.while_loop(cond_fun, body_fun, init_carry)
     aux = {}
-    # print(result)
-    # print(modified)
     return result, aux
 
 
@@ -119,11 +120,12 @@ def _newton_method_manual_parallel(
     y0,
     t,
     dt,
+    args=(),
     modified=False,
     tol=1e-5,
     max_iter=100,
     order=2,
-    args=()
+    
 ):
     r"""
     Newton's method for solving the implicit equations arising from the Crank - Nicolson method for ordinary differential equations (ODEs).
@@ -148,7 +150,7 @@ def _newton_method_manual_parallel(
         max_iter : int, optional
             Maximum number of iterations. Default is 100.
         order : int, optional
-            Order of the integration method. If order = 1, use explicit Euler. If order = 2, use Crank - Nicolson.
+            Order of the integration method. If order = 1, use implicit Euler. If order = 2, use Crank - Nicolson.
         args : tuple, optional
             Additional arguments passed to the function f.
 
@@ -158,6 +160,7 @@ def _newton_method_manual_parallel(
     """
 
     def g(t, y, *args):
+        #jax.debug.print("arg ={a}", a = args)
         if order == 1:
             return y - y0 - dt * f(t + dt, y, *args)[0]
         elif order == 2:
@@ -167,18 +170,19 @@ def _newton_method_manual_parallel(
 
     def cond_fun(carry):
         i, _, cond = carry
-        return u.math.logical_and(i < max_iter, u.math.logical_not(cond))
+        return u.math.logical_and(i < max_iter,cond)
 
     def body_fun(carry):
-        i, y1, _, _ = carry
+        i, y1, _= carry
         # df: [*pop_size, n_compartment, M]
         # A: [*pop_size, n_compartment, M, M]
         A, df = _jacrev_last_dim(lambda y: g(t, y, *args), y1)
+
         shape = df.shape
         # df: [n_neuron * n_compartment, M]
         # A: [n_neuron * n_compartment, M, M]
-        A = A.reshape(-1, shape[-2], shape[-1])
-        df = df.reshape(-1, shape[-1])
+        A = A.reshape((A.shape[0] * A.shape[1],) + A.shape[2:])
+        df = df.reshape((df.shape[0] * df.shape[1],) + df.shape[2:])
 
         # y1: [n_neuron * n_compartment, M]
         condition = u.math.alltrue(
@@ -200,11 +204,11 @@ def _newton_method_manual_parallel(
 
     dt = u.get_magnitude(dt)
     t = u.get_magnitude(t)
-    init_guess = y0  # + dt*f(t, y0, *args)[0]
+    init_guess = y0  + dt*f(t, y0, *args)[0]
     init_carry = (0, init_guess, True)
     n, result, _ = jax.lax.while_loop(
         cond_fun,
-        body_fun_modified if modified else body_fun,
+        body_fun,
         init_carry
     )
     aux = {}
@@ -246,11 +250,21 @@ def _implicit_euler_for_axial_current(A, y0, dt):
         y1 : ndarray
             Solution array at the next time step, shape (n,).
     """
-    # dt = u.get_magnitude(dt)
     n = y0.shape[-1]
-    I = u.math.eye(n)
-    M = I - dt * A
-    y1 = u.math.linalg.solve(M, y0)
+    I = u.math.eye(n) * u.get_unit(dt * A)
+    lhs = I - dt * A
+    rhs = y0
+    y1 = u.math.linalg.solve(lhs, rhs)
+
+    '''
+    # residual
+    residual = rhs - lhs @ y1
+    residual_norm = jnp.linalg.norm(u.get_magnitude(residual))
+    jax.debug.print('Residual norm = {a}', a = residual_norm)
+    jax.debug.print('Relative error = {a}', a = relative_error)
+    cond = jnp.linalg.cond(u.get_magnitude(lhs))
+    jax.debug.print('cond = {a}', a = cond)
+    '''
 
     return y1
 
@@ -291,12 +305,20 @@ def _crank_nicolson_for_axial_current(A, y0, dt):
     """
     n = y0.shape[-1]
     I = u.math.eye(n) * u.get_unit(0.5 * dt * A)
-
     lhs = I - 0.5 * dt * A
-    rhs = (I + 0.5 * dt * A) @ y0
+    rhs = u.math.dot(I + 0.5 * dt * A, y0)
     y1 = u.math.linalg.solve(lhs, rhs)
-    return y1
 
+    '''
+    # residual
+    residual = rhs - lhs @ y1
+    residual_norm = jnp.linalg.norm(u.get_magnitude(residual))
+    jax.debug.print('Residual norm = {a}', a = residual_norm)
+    jax.debug.print('Relative error = {a}', a = relative_error)
+    cond = jnp.linalg.cond(u.get_magnitude(lhs))
+    jax.debug.print('cond = {a}', a = cond)
+    '''
+    return y1
 
 @set_module_as('braincell')
 def implicit_euler_step(
@@ -340,27 +362,23 @@ def construct_A(target):
     A_matrix (array): The constructed matrix A for the axial current.
     """
     with jax.ensure_compile_time_eval():
+        ## load the param
         n_compartment = target.n_compartment
         connection = u.math.array(target.connection)
         cm = target.cm
         A = target.A
         R_axial = target.resistances
-
         pre_ids, post_ids = connection[:, 0], connection[:, 1]
 
+        ## create the A_matrix
         adj_matrix = u.math.zeros((n_compartment, n_compartment)).at[pre_ids, post_ids].set(1)
         R_matrix = u.math.zeros((n_compartment, n_compartment)).at[pre_ids, post_ids].set(
             1 / u.get_magnitude(R_axial)) / u.get_unit(R_axial)
         adj_matrix = adj_matrix + adj_matrix.T
         R_matrix = R_matrix + R_matrix.T
-
-        A_matrix = R_matrix / (cm * A[:, u.math.newaxis])
+        cm_A = cm * A
+        A_matrix = R_matrix / (cm_A[:, u.math.newaxis])
         A_matrix = A_matrix.at[jnp.diag_indices(n_compartment)].set(-u.math.sum(A_matrix, axis=1))
-
-        '''
-        R_matrix = coo_matrix((R_axial, (pre_ids, post_ids)), shape = (n_compartment, n_compartment))
-        A_matrix = coo_matrix(A_matrix)
-        '''
 
     return A_matrix
 
@@ -389,20 +407,21 @@ def splitting_step(
     from braincell.neuron.multi_compartment import MultiCompartment
 
     if isinstance(target, MultiCompartment):
-        # first step, extracting the axial current matrix and solve AV_{n+1} = b(V_n) with Crank–Nicolson method
+        
         def solve_axial():
             dt = brainstate.environ.get_dt()
             V_n = target.V.value
             A_matrix = construct_A(target)
             target.V.value = _crank_nicolson_for_axial_current(A_matrix, V_n, dt)
-
+        '''
         for _ in range(len(target.pop_size)):
             integral = brainstate.augment.vmap(solve_axial, in_states=target.states())
         integral()
-
+        '''
         # second step
 
         with brainstate.environ.context(compute_axial_current=False):
+            '''
             apply_standard_solver_step(
                 _newton_method_manual_parallel,
                 target,
@@ -410,51 +429,28 @@ def splitting_step(
                 *args,
                 merging_method='stack'
             )
+            '''
+            rk4_step(
+                target,
+                t,
+                *args,
+            )
+            '''
+            exp_euler_step(
+                target,
+                t,
+                *args,
+            )
+            '''
 
-            # ralston4_step(
-            #     target,
-            #     t,
-            #     *args,
-            # )
-
-
+        
+        for _ in range(len(target.pop_size)):
+            integral = brainstate.augment.vmap(solve_axial, in_states=target.states())
+        integral()
+        
     else:
         apply_standard_solver_step(
             _newton_method, target, t, *args
         )
 
 
-def _jacrev_last_dim(
-    fn: Callable[[...], jax.Array],
-    hid_vals: jax.Array,
-) -> Tuple[jax.Array, jax.Array]:
-    """
-    Compute the Jacobian of a function with respect to its last dimension.
-
-    This function calculates the Jacobian matrix of the given function 'fn'
-    with respect to the last dimension of the input 'hid_vals'. It uses
-    JAX's vector-Jacobian product (vjp) and vmap for efficient computation.
-
-    Args:
-        fn (Callable[[...], jax.Array]): The function for which to compute
-            the Jacobian. It should take a JAX array as input and return
-            a JAX array.
-        hid_vals (jax.Array): The input values for which to compute the
-            Jacobian. The last dimension is considered as the dimension
-            of interest.
-
-    Returns:
-        jax.Array: The Jacobian matrix. Its shape is (*varshape, num_state, num_state),
-        where varshape is the shape of the input excluding the last dimension,
-        and num_state is the size of the last dimension.
-
-    Raises:
-        AssertionError: If the number of input and output states are not the same.
-    """
-    new_hid_vals, f_vjp = jax.vjp(fn, hid_vals)
-    num_state = new_hid_vals.shape[-1]
-    varshape = new_hid_vals.shape[:-1]
-    assert num_state == hid_vals.shape[-1], 'Error: the number of input/output states should be the same.'
-    g_primals = u.math.broadcast_to(u.math.eye(num_state), (*varshape, num_state, num_state))
-    jac = jax.vmap(f_vjp, in_axes=-2, out_axes=-2)(g_primals)
-    return jac[0], new_hid_vals
